Commented and cleaned TCPClient
So I just finished taking a sweep through this class. I commented just about everything I could in order to help you guys read through it. I also marked any important parts that I could think of for now with a /***************************** comment thing. These are some of the parts you guys will be merging into so maybe have a look at those. I'll be doing this for all the classes.